+++ R/markdown-link.R
@@ -137,7 +137,7 @@ parse_link <- function(destination, contents) {
     )
 
   } else {
-    contents <- gsub("%", "\\\\%", mdxml_link_text(contents))
+    contents <- mdxml_link_text(contents)
 
     list(
       paste0(